+++ app/src/main/java/com/jianfanjia/cn/bean/DesignerInfo.java
@@ -56,9 +56,9 @@ public class DesignerInfo implements Serializable {
 
     private int dec_fee_half;
 
-    private int respond_speed;
+    private float respond_speed;
 
-    private int service_attitude;
+    private float service_attitude;
 
     private List<String> dec_house_types;
 
@@ -339,19 +339,19 @@ public class DesignerInfo implements Serializable {
         return is_block;
     }
 
-    public int getRespond_speed() {
+    public float getRespond_speed() {
         return respond_speed;
     }
 
-    public void setRespond_speed(int respond_speed) {
+    public void setRespond_speed(float respond_speed) {
         this.respond_speed = respond_speed;
     }
 
-    public int getService_attitude() {
+    public float getService_attitude() {
         return service_attitude;
     }
 
-    public void setService_attitude(int service_attitude) {
+    public void setService_attitude(float service_attitude) {
         this.service_attitude = service_attitude;
     }
 }
